Fix authentication (#6)
Authentication state was by default false. This went well on app
startup because the routing also depended on the home assistant
url but when the url was empty and filled again the user was
redirected to the authenticated route again. In theory a
user could therefore get access to someone else's dashboard.